test: reuse generic proxy e2e steps
